+++ vantage6-server/vantage6/server/resource/role.py
@@ -9,12 +9,15 @@ from sqlalchemy import or_
 
 from vantage6.server import db
 from vantage6.server.resource import (
+    get_org_ids_from_collabs,
     with_user,
     ServicesResources
 )
 from vantage6.common import logger_name
 from vantage6.server.permission import (
-    PermissionManager
+    PermissionManager,
+    RuleCollection,
+    Operation as P,
 )
 from vantage6.server.model.rule import Operation, Scope
 from vantage6.server.resource.common._schema import RoleSchema, RuleSchema
@@ -86,20 +89,29 @@ def permissions(permissions: PermissionManager) -> None:
     add = permissions.appender(module_name)
     add(scope=Scope.GLOBAL, operation=Operation.VIEW,
         description="View any role")
+    add(scope=Scope.COLLABORATION, operation=Operation.VIEW,
+        description="View any role in your collaborations")
     add(scope=Scope.ORGANIZATION, operation=Operation.VIEW,
         description="View the roles of your organization")
     add(scope=Scope.GLOBAL, operation=Operation.CREATE,
         description="Create role for any organization")
+    add(scope=Scope.COLLABORATION, operation=Operation.CREATE,
+        description="Create role for any organization in your collaborations")
     add(scope=Scope.ORGANIZATION, operation=Operation.CREATE,
         description="Create role for your organization")
     add(scope=Scope.GLOBAL, operation=Operation.EDIT,
         description="Edit any role")
+    add(scope=Scope.COLLABORATION, operation=Operation.EDIT,
+        description="Edit any role in your collaborations")
     add(scope=Scope.ORGANIZATION, operation=Operation.EDIT,
         description="Edit a role from your organization")
     add(scope=Scope.GLOBAL, operation=Operation.DELETE,
-        description="Delete any organization")
+        description="Delete a role from any organization")
+    add(scope=Scope.COLLABORATION, operation=Operation.DELETE,
+        description="Delete a role from any organization in your "
+                    "collaborations")
     add(scope=Scope.ORGANIZATION, operation=Operation.DELETE,
-        description="Delete your organization")
+        description="Delete a role from your organization")
 
 
 # -----------------------------------------------------------------------------
@@ -113,7 +125,7 @@ class RoleBase(ServicesResources):
 
     def __init__(self, socketio, mail, api, permissions, config):
         super().__init__(socketio, mail, api, permissions, config)
-        self.r = getattr(self.permissions, module_name)
+        self.r: RuleCollection = getattr(self.permissions, module_name)
 
 
 class Roles(RoleBase):
@@ -133,6 +145,8 @@ class Roles(RoleBase):
             Description|\n
             |--|--|--|--|--|--|\n
             |Role|Global|View|❌|❌|View all roles|\n
+            |Role|Collaboration|View|❌|❌|View all roles in your
+            collaborations|\n
             |Role|Organization|View|❌|❌|View roles that are part of your
             organization|\n
 
@@ -164,6 +178,11 @@ class Roles(RoleBase):
                 items:
                   type: integer
                   description: Organization id of which you want to get roles
+            - in: query
+              name: collaboration_id
+              schema:
+              type: integer
+              description: Collaboration id
             - in: query
               name: rule_id
               schema:
@@ -221,10 +240,30 @@ class Roles(RoleBase):
             if 'include_root' in args and args['include_root']:
                 q = q.filter(or_(
                     db.Role.organization_id.in_(org_filters),
-                    db.Role.organization_id == None
+                    db.Role.organization_id.is_(None)
                 ))
             else:
                 q = q.filter(db.Role.organization_id.in_(org_filters))
+
+        # filter by collaboration id
+        if 'collaboration_id' in args:
+            if not self.r.can_for_collaboration(
+                P.VIEW, args['collaboration_id'],
+                self.obtain_auth_collaborations()
+            ):
+                return {
+                    'msg': 'You lack the permission view all roles from '
+                    f'collaboration {args["collaboration_id"]}!'
+                }, HTTPStatus.UNAUTHORIZED
+            org_ids = get_org_ids_from_collabs(g.user,
+                                               args['collaboration_id'])
+            if 'include_root' in args and args['include_root']:
+                q = q.filter(or_(
+                    db.Role.organization_id.in_(org_ids),
+                    db.Role.organization_id.is_(None)
+                ))
+            else:
+                q = q.filter(db.Role.organization_id.in_(org_ids))
 
         # filter by one or more names or descriptions
         for param in ['name', 'description']:
@@ -252,7 +291,7 @@ class Roles(RoleBase):
                 q = q.filter(or_(
                         db.Role.organization_id == auth_org_id,
                         db.Role.id.in_(own_role_ids),
-                        db.Role.organization_id == None
+                        db.Role.organization_id.is_(None)
                     ))
             else:
                 # allow users without permission to view only their own roles
@@ -280,6 +319,8 @@ class Roles(RoleBase):
           Description|\n
           |--|--|--|--|--|--|\n
           |Role|Global|Create|❌|❌|Create a role for any organization|\n
+          |Role|Collaboration|Create|❌|❌|Create a role for organization in
+          your collaborations|\n
           |Role|Organization|Create|❌|❌|Create a role for your organization|\n
 
           Accessible to users.
@@ -361,14 +402,12 @@ class Roles(RoleBase):
                     'exist!'}, HTTPStatus.NOT_FOUND
 
         # check if user is allowed to create this role
-        if (not self.r.c_glo.can() and
-                organization_id != g.user.organization_id):
+        if not self.r.can_by_org(
+            P.CREATE, organization_id, g.user.organization
+        ):
             return {
-                'msg': 'You cannot create roles for other organizations!'
+                'msg': 'You cannot create a role for this organization!'
             }, HTTPStatus.UNAUTHORIZED
-        elif not self.r.c_glo.can() and not self.r.c_org.can():
-            return {'msg': 'You lack the permission to create roles!'}, \
-                HTTPStatus.UNAUTHORIZED
 
         # create the actual role
         role = db.Role(name=data["name"], description=data["description"],
@@ -392,6 +431,8 @@ class Role(RoleBase):
           Description|\n
           |--|--|--|--|--|--|\n
           |Role|Global|View|❌|❌|View all roles|\n
+          |Role|Collaboration|View|❌|❌|View all roles for your
+          collaborations|\n
           |Role|Organization|View|❌|❌|View roles that are part of your
           organization|\n
 
@@ -424,11 +465,13 @@ class Role(RoleBase):
                 HTTPStatus.NOT_FOUND
 
         # check permissions. A user can always view their own roles
-        if not (self.r.v_glo.can() or role in g.user.roles):
-            if not (self.r.v_org.can()
-                    and role.organization == g.user.organization):
-                return {"msg": "You do not have permission to view this."},\
-                     HTTPStatus.UNAUTHORIZED
+        if not (
+            self.r.can_by_org(P.VIEW, role.organization_id,
+                              g.user.organization) or
+            role in g.user.roles
+        ):
+            return {"msg": "You do not have permission to view this."},\
+                    HTTPStatus.UNAUTHORIZED
 
         return role_schema.dump(role, many=False).data, HTTPStatus.OK
 
@@ -444,6 +487,8 @@ class Role(RoleBase):
           Description|\n
           |--|--|--|--|--|--|\n
           |Role|Global|Edit|❌|❌|Update any role|\n
+          |Role|Collaboration|Edit|❌|❌|Update any role in your
+          collaborations|\n
           |Role|Organization|Edit|❌|❌|Update a role from your organization|\n
 
           Accessible to users.
@@ -509,13 +554,10 @@ class Role(RoleBase):
             }, HTTPStatus.BAD_REQUEST
 
         # check permission of the user
-        if not self.r.e_glo.can():
-            if not self.r.e_org.can():
-                return {'msg': 'You do not have permission to edit roles!'}, \
-                    HTTPStatus.UNAUTHORIZED
-            elif g.user.organization_id != role.organization.id:
-                return {'msg': 'You can\'t edit roles from another '
-                        'organization'}, HTTPStatus.UNAUTHORIZED
+        if not self.r.can_by_org(P.EDIT, role.organization_id,
+                                 g.user.organization):
+            return {'msg': 'You do not have permission to edit this role!'}, \
+                HTTPStatus.UNAUTHORIZED
 
         # process patch
         if 'name' in data:
@@ -551,6 +593,8 @@ class Role(RoleBase):
           Description|\n
           |--|--|--|--|--|--|\n
           |Role|Global|Delete|❌|❌|Delete any role|\n
+          |Role|Collaboration|Delete|❌|❌|Delete any role in your
+          collaborations|\n
           |Role|Organization|Delete|❌|❌|Delete a role in your organization|\n
 
           Accessible to users.
@@ -567,6 +611,8 @@ class Role(RoleBase):
         responses:
           200:
             description: Ok
+          400:
+            description: Cannot delete default roles
           401:
             description: Unauthorized
           404:
@@ -582,13 +628,16 @@ class Role(RoleBase):
             return {"msg": f"Role with id={id} not found."}, \
                 HTTPStatus.NOT_FOUND
 
-        if not self.r.d_glo.can():
-            if not self.r.d_org.can():
-                return {'msg': 'You do not have permission to delete roles!'},\
-                    HTTPStatus.UNAUTHORIZED
-            elif role.organization.id != g.user.organization.id:
-                return {'msg': 'You can\'t delete a role from another '
-                        'organization'}, HTTPStatus.UNAUTHORIZED
+        if role.name in [role for role in DefaultRole]:
+            return {
+                "msg": f"This role ('{role.name}') is a default role. Default"
+                       " roles cannot be deleted."
+            }, HTTPStatus.BAD_REQUEST
+
+        if not self.r.can_by_org(P.DELETE, role.organization_id,
+                                 g.user.organization):
+            return {'msg': 'You do not have permission to delete this role!'},\
+                HTTPStatus.UNAUTHORIZED
 
         role.delete()
